\name{isolation_forest_etiquetar}
\alias{isolation_forest_etiquetar}
\title{Detección por Isolation Forest}
\description{Detecta atípicos usando bosques de aislamiento (paquete isotree).}
\usage{isolation_forest_etiquetar(data, contaminacion = 0.01, n_arboles = 200, max_muestras = 'auto', semilla = NULL)}
\details{}
\keyword{package}
